break out more modules and add loader
